Add files via upload
This is a script to take data from the FastF1 api from Japan 2024, as well as takes data from the 2025 races so far, and tries to predict the Q3 times of each driver. There are also team and individual driver factors applied as well. 

Citations:
Github user mar-antaya made most of this code, I am mostly using it for personal testing and python library exploration.